+++ src/main/java/com/gznznzjsn/outstagram/service/impl/driver/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Implementations of services with Neo4j Java Driver.
+ */
+package com.gznznzjsn.outstagram.service.impl.driver;
