--- src/main/java/com/gznznzjsn/outstagram/service/impl/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implementations of services.
- */
-package com.gznznzjsn.outstagram.service.impl;
